requirements_new.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Pillow==10.0.1
Werkzeug==2.3.7
